SIG-1534 split up v1 serializers move validation
