+++ api/app/signals/apps/api/v1/serializers/nested/attachment.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+from signals.apps.signals.models import Attachment
+
+
+class _NestedAttachmentModelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Attachment
+        fields = (
+            'file',
+            'created_at',
+            'is_image',
+        )
