+++ api/app/signals/apps/api/v1/serializers/nested/location.py
@@ -1,0 +1,29 @@
+from rest_framework import serializers
+
+from signals.apps.api.generics.validators import NearAmsterdamValidatorMixin
+from signals.apps.signals.models import Location
+
+
+class _NestedLocationModelSerializer(NearAmsterdamValidatorMixin, serializers.ModelSerializer):
+    class Meta:
+        model = Location
+        geo_field = 'geometrie'
+        fields = (
+            'id',
+            'stadsdeel',
+            'buurt_code',
+            'address',
+            'address_text',
+            'geometrie',
+            'extra_properties',
+            'created_by',
+            'bag_validated',
+        )
+        read_only_fields = (
+            'id',
+            'created_by',
+            'bag_validated',
+        )
+        extra_kwargs = {
+            'id': {'label': 'ID', },
+        }
