+++ api/app/signals/apps/api/v1/serializers/nested/note.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from signals.apps.signals.models import Note
+
+
+class _NestedNoteModelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Note
+        fields = (
+            'text',
+            'created_by',
+        )
+        read_only_fields = (
+            'created_by',
+        )
