+++ api/app/signals/apps/api/v1/serializers/nested/reporter.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+
+from signals.apps.signals.models import Reporter
+
+
+class _NestedReporterModelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Reporter
+        fields = (
+            'email',
+            'phone',
+        )
