+++ api/app/signals/apps/api/v1/serializers/nested/status.py
@@ -1,0 +1,54 @@
+from rest_framework import serializers
+from rest_framework.exceptions import PermissionDenied
+
+from signals.apps.signals import workflow
+from signals.apps.signals.models import Status
+
+
+class _NestedStatusModelSerializer(serializers.ModelSerializer):
+    state_display = serializers.CharField(source='get_state_display', read_only=True)
+
+    class Meta:
+        model = Status
+        fields = (
+            'text',
+            'user',
+            'state',
+            'state_display',
+            'target_api',
+            'extra_properties',
+            'created_at',
+        )
+        read_only_fields = (
+            'created_at',
+            'user',
+        )
+
+    def validate(self, attrs):
+        if (attrs['state'] == workflow.TE_VERZENDEN
+                and attrs.get('target_api') == Status.TARGET_API_SIGMAX):
+
+            request = self.context.get('request')
+            if request and not request.user.has_perm('signals.push_to_sigmax'):
+                raise PermissionDenied({
+                    'state': "You don't have permissions to push to Sigmax/CityControl."
+                })
+
+        return super(_NestedStatusModelSerializer, self).validate(attrs=attrs)
+
+
+class _NestedPublicStatusModelSerializer(serializers.ModelSerializer):
+    state_display = serializers.CharField(source='get_state_display', read_only=True)
+
+    class Meta:
+        model = Status
+        fields = (
+            'id',
+            'state',
+            'state_display',
+        )
+        read_only_fields = (
+            'id',
+            'state',
+            'state_display',
+        )
