+++ api/app/tests/apps/api/test_api_v1_endpoints.py
@@ -12,10 +12,7 @@ from freezegun import freeze_time
 from rest_framework.reverse import reverse
 
 from signals import API_VERSIONS
-from signals.apps.api.address.validation import (
-    AddressValidationUnavailableException,
-    NoResultsException
-)
+from signals.apps.api.v1.validation import AddressValidationUnavailableException, NoResultsException
 from signals.apps.feedback.models import Feedback
 from signals.apps.signals import workflow
 from signals.apps.signals.models import Attachment, Category, History, Signal
@@ -548,7 +545,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
 
     # -- write tests --
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict",
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict",
            side_effect=AddressValidationUnavailableException)  # Skip address validation
     def test_create_initial(self, validate_address_dict):
         # Authenticate, load fixture and add relevant main and sub category.
@@ -586,7 +583,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
 
         self.assertEqual(400, response.status_code)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict",
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict",
            side_effect=NoResultsException)
     def test_create_initial_invalid_location(self, validate_address_dict):
         """ Tests that a 400 is returned when an invalid location is provided """
@@ -596,7 +593,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
 
         self.assertEqual(response.status_code, 400)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict")
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict")
     def test_create_initial_valid_location(self, validate_address_dict):
         """ Tests that bag_validated is set to True when a valid location is provided and that
         the address is replaced with the suggested address. The original address should be saved
@@ -627,7 +624,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = self.client.get(new_url).json()
         self.assertJsonSchema(self.retrieve_signal_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict")
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict")
     def test_create_initial_valid_location_but_no_address(self, validate_address_dict):
         """Tests that a Signal can be created when loccation has no known address but
         coordinates are known."""
@@ -650,7 +647,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = self.client.get(new_url).json()
         self.assertJsonSchema(self.retrieve_signal_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict",
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict",
            side_effect=AddressValidationUnavailableException)
     def test_create_initial_address_validation_unavailable(self, validate_address_dict):
         """ Tests that the signal is created even though the address validation service is
@@ -673,7 +670,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = self.client.get(new_url).json()
         self.assertJsonSchema(self.retrieve_signal_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict",
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict",
            side_effect=AddressValidationUnavailableException)
     def test_create_initial_try_update_bag_validated(self, validate_address_dict):
         """ Tests that the bag_validated field cannot be set manually, and that the address
@@ -700,7 +697,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = self.client.get(new_url).json()
         self.assertJsonSchema(self.retrieve_signal_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict",
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict",
            side_effect=AddressValidationUnavailableException)  # Skip address validation
     def test_create_initial_and_upload_image(self, validate_address_dict):
         # Authenticate, load fixture.
@@ -730,7 +727,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = self.client.get(new_url).json()
         self.assertJsonSchema(self.retrieve_signal_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict")
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict")
     def test_update_location(self, validate_address_dict):
         # Partial update to update the location, all interaction via API.
         self.client.force_authenticate(user=self.sia_read_write_user)
@@ -770,7 +767,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = response.json()
         self.assertJsonSchema(self.list_history_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict")
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict")
     def test_update_location_no_address(self, validate_address_dict):
         # Partial update to update the location, all interaction via API.
         # SIA must also allow location updates without known address but with
@@ -812,7 +809,7 @@ class TestPrivateSignalViewSet(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
         response_json = response.json()
         self.assertJsonSchema(self.list_history_schema, response_json)
 
-    @patch("signals.apps.api.address.validation.AddressValidation.validate_address_dict")
+    @patch("signals.apps.api.v1.validation.AddressValidation.validate_address_dict")
     def test_update_location_no_coordinates(self, validate_address_dict):
         # Partial update to update the location, all interaction via API.
         # SIA must also allow location updates without known address but with
